+++ CHANGELOG.md
@@ -3,5 +3,5 @@ All notable changes to this project are documented in this file.
 
 ## 12.x.x
 
-* `BitbucketClient` now needs a WSClient to be built, more info on README depending on the Play version that you are using
+* `BitbucketClient` for Play 2.7+ now needs a akka.stream.Materializer to be built
 * Removed OAuth1 authentication method